User now has full access to integrate any type of card to process the payment via Stripe.

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -22,6 +22,7 @@ export class CheckoutPaymentComponent implements OnInit {
   cardNumber?: StripeCardNumberElement;
   cardExpiry?: StripeCardExpiryElement;
   cardCvc?: StripeCardCvcElement;
+  cardErrors: any;
 
   constructor(private basketService: BasketService, private checkoutService: CheckoutService,
       private toastr: ToastrService, private router: Router
@@ -35,12 +36,24 @@ export class CheckoutPaymentComponent implements OnInit {
       if(elements) {
         this.cardNumber = elements.create('cardNumber');
         this.cardNumber.mount(this.cardNumberElement?.nativeElement);
+        this.cardNumber.on('change', event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
         
         this.cardExpiry = elements.create('cardExpiry');
         this.cardExpiry.mount(this.cardExpiryElement?.nativeElement);
+        this.cardExpiry.on('change', event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
 
         this.cardCvc = elements.create('cardCvc');
         this.cardCvc.mount(this.cardCvcElement?.nativeElement);
+        this.cardCvc.on('change', event => {
+          if(event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
       }
     })
   }
@@ -54,9 +67,21 @@ export class CheckoutPaymentComponent implements OnInit {
     this.checkoutService.createOrder(orderToCreate).subscribe({
       next: order => {
         this.toastr.success('Order created successfully');
-        this.basketService.deleteLocalBasket();
-        const navigaTionExtras: NavigationExtras = {state: order};
-        this.router.navigate(['checkout/success'], navigaTionExtras);
+        this.stripe?.confirmCardPayment(basket.clientSecret!, {
+          payment_method: {
+            card: this.cardNumber!,
+            billing_details: {
+              name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+            }
+          }
+        }).then(result => {
+          console.log(result)
+          if(result.paymentIntent) {   
+            this.basketService.deleteLocalBasket();
+            const navigaTionExtras: NavigationExtras = {state: order};
+            this.router.navigate(['checkout/success'], navigaTionExtras);
+          }
+        })
       }
     }) 
   }
